A lot of changes vol2

+++ src/main/java/com/nick1est/proconnectx/dto/LightweightServiceDto.java
@@ -1,12 +1,12 @@
 package com.nick1est.proconnectx.dto;
 
 import com.nick1est.proconnectx.dao.CategoryType;
-import com.nick1est.proconnectx.dto.employer.registration.FileDto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -24,7 +24,7 @@ public class LightweightServiceDto {
     @NotNull
     private Integer ratingCount;
     @NotNull
-    private Double price;
+    private BigDecimal price;
     @NotNull
     private LightWeightFreelancerDto freelancer;
     @NotEmpty
